requirements.txt
blurhash==1.1.4
certifi==2022.12.7
charset-normalizer==3.0.1
decorator==5.1.1
future==0.18.3
idna==3.4
Mastodon.py==1.8.0
oauthlib==3.2.2
python-dateutil==2.8.2
python-magic==0.4.27
python-twitter==3.5
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.14
